abstracted the widget
